Task2/requirements.txt
openai
python-dotenv
asyncpg